+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * BC: add OpenType styles
 * BC: self links are removed
+* BC: url() in CSS is now relative to the CSS file, images should be included in the same directory
 
 # 2.0.0